6 / 27

+++ Assets/01Scripts/Interface/IDamaged.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface IDamaged
 {
-    void TakeDamage(float amount);
+    void TakeDamage(int amount);
 }
